Conflicts:
	Detour/Include/DetourNode.h
	Detour/Source/DetourNavMeshQuery.cpp
	DetourCrowd/Include/DetourCrowd.h
	DetourCrowd/Source/DetourCrowd.cpp
	DetourCrowd/Source/DetourObstacleAvoidance.cpp
	Recast/Source/RecastMeshDetail.cpp

+++ DetourCrowd/Source/DetourObstacleAvoidance.cpp
@@ -331,7 +331,7 @@ float dtObstacleAvoidanceQuery::processSample(const float* vcand, const float cs
 	// find the threshold hit time to bail out based on the early out penalty
 	// (see how the penalty is calculated below to understnad)
 	float minPen = minPenalty - vpen - vcpen;
-	float tThresold = (float)(((double)m_params.weightToi/(double)minPen - 0.1) * (double)m_params.horizTime);
+	float tThresold = ((double)m_params.weightToi/(double)minPen - 0.1) * (double)m_params.horizTime;
 	if (tThresold - m_params.horizTime > -FLT_EPSILON)
 		return minPenalty; // already too much
 
@@ -538,7 +538,7 @@ int dtObstacleAvoidanceQuery::sampleVelocityAdaptive(const float* pos, const flo
 	dtVcopy(ddir, dvel);
 	dtNormalize2D(ddir);
 	dtRorate2D (ddir+3, ddir, da*0.5f); // rotated by da/2
-
+	
 	// Always add sample at zero
 	pat[npat*2+0] = 0;
 	pat[npat*2+1] = 0;
@@ -574,7 +574,6 @@ int dtObstacleAvoidanceQuery::sampleVelocityAdaptive(const float* pos, const flo
 			npat++;
 		}
 	}
-
 
 	// Start sampling.
 	float cr = vmax * (1.0f - m_params.velBias);
